Revamp /tourladder
Make tourladder display top 10 in default and show help message on bottom
about showing the entire ladder or certain amount of users on the ladder.
This is to make it more user friendly and prevent a giant flood cause by
mass amount of user on the tour ladder. Also center the tourladder for
emphasis.